Add: Kafka 기능 & Zipkin 추가

See merge request s09-final/S09P31A204!11

+++ journey-service/src/main/java/com/ssafy/journeymate/journeyservice/client/CategoryServiceClient.java
@@ -7,13 +7,13 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
-@FeignClient(name="category-service")
+@FeignClient(name = "category-service")
 public interface CategoryServiceClient {
 
     @GetMapping("/category-service/{categoryId}")
-    CategoryGetRes getCategory(@PathVariable int categoryId);
+    CategoryGetRes getCategory(@PathVariable Long categoryId);
 
     @GetMapping("/category-service/{categoryId}/items")
-    List<ItemGetRes> getCategoryItems(@PathVariable int categoryId);
+    List<ItemGetRes> getCategoryItems(@PathVariable Long categoryId);
 
 }
